% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{plot_manhattan}
\usage{
plot_manhattan(
  df,
  chr_col = "CHR",
  bp_col = "BP",
  p_col = "P",
  sig_line = 5e-08,
  title = "Manhattan Plot"
)
}
\arguments{
\item{df}{A data.frame containing at least the following columns:
- CHR: chromosome identifier
- BP: base-pair position
- P: p-value}

\item{chr_col}{Name of the chromosome column (string). Default: "CHR"}

\item{bp_col}{Name of the base-pair position column (string). Default: "BP"}

\item{p_col}{Name of the p-value column (string). Default: "P"}

\item{sig_line}{Significance threshold for p-values. Default: 5e-8}

\item{title}{Plot title. Default: "Manhattan Plot"}
}
\value{
A ggplot2 object representing the Manhattan plot
}
\description{
Generate a Manhattan plot from a PLINK assoc.logistic data.frame
}
